requirements.txt
django-embed-video
python-memcached
django-memcache-status==1.3